remove prefetcher dependency on router
